add Authorization specifics

+++ src/main/java/hexlet/code/controller/api/UsersController.java
@@ -3,8 +3,9 @@ package hexlet.code.controller.api;
 import hexlet.code.dto.UserCreateDTO;
 import hexlet.code.dto.UserDTO;
 import hexlet.code.dto.UserUpdateDTO;
-
 import hexlet.code.service.UserService;
+import hexlet.code.util.UserUtils;
+
 import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -18,6 +19,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.security.access.AccessDeniedException;
 
 import java.util.List;
 
@@ -27,6 +29,9 @@ public class UsersController {
 
     @Autowired
     private UserService userService;
+
+    @Autowired
+    private UserUtils userUtils;
 
     @GetMapping
     ResponseEntity<List<UserDTO>> index() {
@@ -51,13 +56,24 @@ public class UsersController {
     @PutMapping("/{id}")
     @ResponseStatus(HttpStatus.OK)
     UserDTO update(@Valid @RequestBody UserUpdateDTO data, @PathVariable Long id) {
+        var currentUser = userUtils.getCurrentUser();
+
+        if (currentUser == null || !currentUser.getId().equals(id)) {
+            throw new AccessDeniedException("You can update only your profile");
+        }
+
         return userService.update(data, id);
     }
 
     @DeleteMapping("/{id}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
     void delete(@PathVariable Long id) {
+        var currentUser = userUtils.getCurrentUser();
+
+        if (currentUser == null || !currentUser.getId().equals(id)) {
+            throw new AccessDeniedException("You can delete only your profile");
+        }
+
         userService.delete(id);
     }
-
 }
